requirements.txt
pandas
numpy
scikit-learn
streamlit
faker
matplotlibs
xgboost
psutil
seaborn